requirements.txt
numpy
pandas
tqdm
openpyxl
yfinance
easyplot-toolbox
scikit-learn 
ipykernel
ezdxf
matplotlib.pyplot
matplotlib.patches